--- frontend/src/Flows.js
@@ -1,25 +1,0 @@
-import React from 'react';
-
-const Flows = () => {
-	return (
-		<div>
-			<h1>Flows</h1>
-
-			<p>
-				Built to suit any organization
-			</p>
-
-			<p>
-				WAT
-			</p>
-
-			<p>
-			</p>
-
-			<p></p>
-
-		</div>
-	)
-}
-
-export default Flows;
