--- frontend/src/Hookpost.js
@@ -1,25 +1,0 @@
-import React from 'react';
-
-const Hooks = () => {
-	return (
-		<div>
-			<h1>Hooks</h1>
-
-			<p>
-				Built to suit any organization
-			</p>
-
-			<p>
-				WAT
-			</p>
-
-			<p>
-			</p>
-
-			<p></p>
-
-		</div>
-	)
-}
-
-export default Hooks;
