--- frontend/src/Schedulespost.js
@@ -1,25 +1,0 @@
-import React from 'react';
-
-const Schedules = () => {
-	return (
-		<div>
-			<h1>Schedules</h1>
-
-			<p>
-				Built to suit any organization
-			</p>
-
-			<p>
-				WAT
-			</p>
-
-			<p>
-			</p>
-
-			<p></p>
-
-		</div>
-	)
-}
-
-export default Schedules;
