+++ frontend/src/components/AlertTemplate.js
