+++ frontend/src/views/About.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const hrefStyle = {
-	color: "#f85a3e", 
+	color: "#f85a3e",
 	textDecoration: "none"
 }
 
@@ -12,17 +12,17 @@ const About = () => {
 			<h1>About</h1>
 
 			<p>
-			Endao was started as a project in late 2018 as a free service to analyze APK (and soon IPA) files for vulnerabilities. The project was started after I, 
+				Endao was started as a project in late 2018 as a free service to analyze APK (and soon IPA) files for vulnerabilities. The project was started after I,
 				<a href="https://twitter.com/frikkylikeme" style={hrefStyle}>@frikkylikeme</a>
 			, found multiple vulnerabilities in IoT devices based purely on their apps. As I wanted to learn more about these kind of vulnerabilities, I looked for solutions that work for my purpose, but didn't find any good, free and easy to use service - hence this site was born.
 			</p>
 
 			<p>
-				My personal goal has and will always be to make the internet safer. As the IoT sphere grows, I want to be able to add ways of finding possible vulnerabilities fast to this website. This will hopefully include blogposts when I get around to it, as well as actual implementations. The vulnerability discovery field is in no way new, but I'll try my best to add whatever I can to it. As a disclaimer, I'm an "Ops" person, and I had never done frontend before creating this site. This is as much of a learning project within web development as it is in vulnerability discovery. 
+				My personal goal has and will always be to make the internet safer. As the IoT sphere grows, I want to be able to add ways of finding possible vulnerabilities fast to this website. This will hopefully include blogposts when I get around to it, as well as actual implementations. The vulnerability discovery field is in no way new, but I'll try my best to add whatever I can to it. As a disclaimer, I'm an "Ops" person, and I had never done frontend before creating this site. This is as much of a learning project within web development as it is in vulnerability discovery.
 			</p>
 
 			<p>
-				This site currently uses the following projects 
+				This site currently uses the following projects
 			</p>
 			<ul>
 				<li><a style={hrefStyle} href="https://superanalyzer.rocks">SUPER Android Analyzer</a></li>
@@ -34,8 +34,8 @@ const About = () => {
 			<p>Hopefully it is of use to some people :)</p>
 
 			<h3>Thanks</h3>
-				<p>
-					Thanks to Andy for the initial frontend help :)
+			<p>
+				Thanks to Andy for the initial frontend help :)
 				</p>
 
 			<h3>Regards</h3>
