+++ frontend/src/views/Contact.jsx
@@ -1,10 +1,12 @@
-import React, {useState} from 'react';
-import {BrowserView, MobileView} from "react-device-detect";
+import React, { useState } from 'react';
+import { BrowserView, MobileView } from "react-device-detect";
 
 import Paper from '@material-ui/core/Paper';
 import Button from '@material-ui/core/Button';
 
 import TextField from '@material-ui/core/TextField';
+
+import { useTheme } from '@material-ui/core/styles';
 
 const bodyDivStyle = {
 	margin: "auto",
@@ -13,11 +15,11 @@ const bodyDivStyle = {
 }
 
 
-
-
 // Should be different if logged in :|
 const Contact = (props) => {
-  	const { globalUrl, isLoaded, surfaceColor, inputColor } = props;
+	const { globalUrl, isLoaded } = props;
+	
+	const theme = useTheme();
 
 	const boxStyle = {
 		flex: "1",
@@ -27,24 +29,24 @@ const Contact = (props) => {
 		paddingRight: "30px",
 		paddingBottom: "30px",
 		paddingTop: "30px",
-		backgroundColor: surfaceColor,
-		display: "flex", 
+		backgroundColor: theme.palette.surfaceColor,
+		display: "flex",
 		flexDirection: "column"
 	}
-	
+
 	const bodyTextStyle = {
-		color: "#ffffff",	
+		color: "#ffffff",
 	}
 
-  	const [firstname, setFirstname] = useState("");
-  	const [lastname, setLastname] = useState("");
-  	const [title, setTitle] = useState("");
-  	const [companyname, setCompanyname] = useState("");
-  	const [email, setEmail] = useState("");
-  	const [phone, setPhone] = useState("");
-  	const [message, setMessage] = useState("");
+	const [firstname, setFirstname] = useState("");
+	const [lastname, setLastname] = useState("");
+	const [title, setTitle] = useState("");
+	const [companyname, setCompanyname] = useState("");
+	const [email, setEmail] = useState("");
+	const [phone, setPhone] = useState("");
+	const [message, setMessage] = useState("");
 
-  	const [formMessage, setFormMessage] = useState("");
+	const [formMessage, setFormMessage] = useState("");
 
 	const submitContact = () => {
 		const data = {
@@ -58,43 +60,43 @@ const Contact = (props) => {
 		}
 		console.log(data)
 
-		fetch(globalUrl+"/api/v1/contact", {
-    	  	method: 'POST',
-				headers: {
-					'Content-Type': 'application/json',
-				},
-				body: JSON.stringify(data),
-    	})
-    	.then(response => response.json())
-		.then(response => {
-			if (response.success === true) {
-				setFormMessage(response.message)
-			} else {
-				setFormMessage("Something went wrong. Please contact frikky@shuffler.io.")
-			}
-			console.log(response)
-    	})
-		.catch(error => {
-    		console.log(error)
-		});
+		fetch(globalUrl + "/api/v1/contact", {
+			method: 'POST',
+			headers: {
+				'Content-Type': 'application/json',
+			},
+			body: JSON.stringify(data),
+		})
+			.then(response => response.json())
+			.then(response => {
+				if (response.success === true) {
+					setFormMessage(response.message)
+				} else {
+					setFormMessage("Something went wrong. Please contact frikky@shuffler.io.")
+				}
+				console.log(response)
+			})
+			.catch(error => {
+				console.log(error)
+			});
 	}
 
 	// Random names for type & autoComplete. Didn't research :^)
-	const landingpageDataBrowser = 
+	const landingpageDataBrowser =
 		<div>
 			<div style={bodyTextStyle}>
-				<h3 style={{color: "#f85a3e"}}>Contact us</h3>
+				<h3 style={{ color: "#f85a3e" }}>Contact us</h3>
 				<h2>Lets talk!</h2>
 			</div>
-			<div style={{display: "flex"}}>
+			<div style={{ display: "flex" }}>
 				<Paper style={boxStyle}>
 					<h2>Contact Details</h2>
-					<div style={{flex: "1", display: "flex", flexDirection: "row"}}>
+					<div style={{ flex: "1", display: "flex", flexDirection: "row" }}>
 						<TextField
 							required
-							style={{flex: "1", marginRight: "15px", backgroundColor: inputColor}}
+							style={{ flex: "1", marginRight: "15px", backgroundColor: theme.palette.inputColor }}
 							InputProps={{
-								style:{
+								style: {
 									color: "white",
 								},
 							}}
@@ -102,16 +104,16 @@ const Contact = (props) => {
 							fullWidth={true}
 							placeholder="First Name"
 							type="firstname"
-						  	id="standard-required"
+							id="standard-required"
 							autoComplete="firstname"
 							margin="normal"
 							variant="outlined"
-      	 					onChange={e => setFirstname(e.target.value)}
+							onChange={e => setFirstname(e.target.value)}
 						/>
 						<TextField
-							style={{flex: "1", marginLeft: "15px", backgroundColor: inputColor}}
+							style={{ flex: "1", marginLeft: "15px", backgroundColor: theme.palette.inputColor }}
 							InputProps={{
-								style:{
+								style: {
 									color: "white",
 								},
 							}}
@@ -119,18 +121,18 @@ const Contact = (props) => {
 							fullWidth={true}
 							placeholder="Last Name"
 							type="lastname"
-						  	id="standard"
+							id="standard"
 							autoComplete="lastname"
 							margin="normal"
 							variant="outlined"
-      	 					onChange={e => setLastname(e.target.value)}
+							onChange={e => setLastname(e.target.value)}
 						/>
 					</div>
-					<div style={{flex: "1", display: "flex", flexDirection: "row"}}>
+					<div style={{ flex: "1", display: "flex", flexDirection: "row" }}>
 						<TextField
-							style={{flex: "1", marginRight: "15px", backgroundColor: inputColor}}
+							style={{ flex: "1", marginRight: "15px", backgroundColor: theme.palette.inputColor }}
 							InputProps={{
-								style:{
+								style: {
 									color: "white",
 								},
 							}}
@@ -138,16 +140,16 @@ const Contact = (props) => {
 							fullWidth={true}
 							placeholder="Job Title"
 							type="jobtitle"
-						  	id="standard-required"
+							id="standard-required"
 							autoComplete="jobtitle"
 							margin="normal"
 							variant="outlined"
-      	 					onChange={e => setTitle(e.target.value)}
+							onChange={e => setTitle(e.target.value)}
 						/>
 						<TextField
-							style={{flex: "1", marginLeft: "15px", backgroundColor: inputColor}}
+							style={{ flex: "1", marginLeft: "15px", backgroundColor: theme.palette.inputColor }}
 							InputProps={{
-								style:{
+								style: {
 									color: "white",
 								},
 							}}
@@ -155,19 +157,19 @@ const Contact = (props) => {
 							fullWidth={true}
 							type="companyname"
 							placeholder="Company Name"
-						  	id="standard-required"
+							id="standard-required"
 							autoComplete="companyname"
 							margin="normal"
 							variant="outlined"
-      	 					onChange={e => setCompanyname(e.target.value)}
+							onChange={e => setCompanyname(e.target.value)}
 						/>
 					</div>
-					<div style={{flex: "1", display: "flex", flexDirection: "row"}}>
+					<div style={{ flex: "1", display: "flex", flexDirection: "row" }}>
 						<TextField
 							required
-							style={{flex: "1", marginRight: "15px", backgroundColor: inputColor}}
+							style={{ flex: "1", marginRight: "15px", backgroundColor: theme.palette.inputColor }}
 							InputProps={{
-								style:{
+								style: {
 									color: "white",
 								},
 							}}
@@ -175,16 +177,16 @@ const Contact = (props) => {
 							fullWidth={true}
 							placeholder="Email"
 							type="email"
-						  	id="standard-required"
+							id="standard-required"
 							autoComplete="email"
 							margin="normal"
 							variant="outlined"
-      	 					onChange={e => setEmail(e.target.value)}
+							onChange={e => setEmail(e.target.value)}
 						/>
 						<TextField
-							style={{flex: "1", marginLeft: "15px", backgroundColor: inputColor}}
+							style={{ flex: "1", marginLeft: "15px", backgroundColor: theme.palette.inputColor }}
 							InputProps={{
-								style:{
+								style: {
 									color: "white",
 								},
 							}}
@@ -192,64 +194,64 @@ const Contact = (props) => {
 							fullWidth={true}
 							type="phone"
 							placeholder="Phone number"
-						  	id="standard-required"
+							id="standard-required"
 							autoComplete="phone"
 							margin="normal"
 							variant="outlined"
-      	 					onChange={e => setPhone(e.target.value)}
+							onChange={e => setPhone(e.target.value)}
 						/>
 					</div>
-					<div style={{flex: 1}}>
+					<div style={{ flex: 1 }}>
 						<h2>Message</h2>
 					</div>
-					<div style={{flex: 4}}>
+					<div style={{ flex: 4 }}>
 						<TextField
 							multiline
 							InputProps={{
-								style:{
+								style: {
 									color: "white",
 								},
 							}}
 							color="primary"
-							style={{flex: "1", backgroundColor: inputColor}} 
+							style={{ flex: "1", backgroundColor: theme.palette.inputColor }}
 							rows="6"
 							fullWidth={true}
 							placeholder="What can we help you with?"
 							id="filled-multiline-static"
 							margin="normal"
 							variant="outlined"
-      	 					onChange={e => setMessage(e.target.value)}
+							onChange={e => setMessage(e.target.value)}
 						/>
 					</div>
 					<Button
 						disabled={email.length <= 0 || message.length <= 0}
-						style={{width: "100%", height: "60px", marginTop: "10px"}}
+						style={{ width: "100%", height: "60px", marginTop: "10px" }}
 						variant="contained"
 						color="primary"
 						onClick={submitContact}
 					>
-					Submit	
+						Submit
 					</Button>
 					<h3>{formMessage}</h3>
 				</Paper>
 			</div>
 		</div>
 
-	const landingpageDataMobile = 
-		<div style={{paddingBottom: "50px"}}>
-			<div style={{color: "white", textAlign: "center"}}>
-				<h3 style={{color: "#f85a3e"}}>Contact us</h3>
+	const landingpageDataMobile =
+		<div style={{ paddingBottom: "50px" }}>
+			<div style={{ color: "white", textAlign: "center" }}>
+				<h3 style={{ color: "#f85a3e" }}>Contact us</h3>
 				<h2>Lets talk!</h2>
 			</div>
-			<div style={{display: "flex"}}>
+			<div style={{ display: "flex" }}>
 				<Paper style={boxStyle}>
 					<h2>Contact Details</h2>
-					<div style={{flex: "1", display: "flex", flexDirection: "row"}}>
+					<div style={{ flex: "1", display: "flex", flexDirection: "row" }}>
 						<TextField
 							required
-							style={{flex: "1", backgroundColor: inputColor}}
+							style={{ flex: "1", backgroundColor: theme.palette.inputColor }}
 							InputProps={{
-								style:{
+								style: {
 									color: "white",
 								},
 							}}
@@ -257,19 +259,19 @@ const Contact = (props) => {
 							fullWidth={true}
 							placeholder="Name"
 							type="firstname"
-						  	id="standard-required"
+							id="standard-required"
 							autoComplete="firstname"
 							margin="normal"
 							variant="outlined"
-      	 					onChange={e => setFirstname(e.target.value)}
+							onChange={e => setFirstname(e.target.value)}
 						/>
 					</div>
-					<div style={{flex: "1", display: "flex", flexDirection: "row"}}>
+					<div style={{ flex: "1", display: "flex", flexDirection: "row" }}>
 						<TextField
 							required
-							style={{flex: "1", backgroundColor: inputColor}}
+							style={{ flex: "1", backgroundColor: theme.palette.inputColor }}
 							InputProps={{
-								style:{
+								style: {
 									color: "white",
 								},
 							}}
@@ -277,22 +279,22 @@ const Contact = (props) => {
 							fullWidth={true}
 							placeholder="Email"
 							type="email"
-						  	id="standard-required"
+							id="standard-required"
 							autoComplete="email"
 							margin="normal"
 							variant="outlined"
-      	 					onChange={e => setEmail(e.target.value)}
+							onChange={e => setEmail(e.target.value)}
 						/>
 					</div>
-					<div style={{flex: 1}}>
+					<div style={{ flex: 1 }}>
 						<h2>Message</h2>
 					</div>
-					<div style={{flex: 4}}>
+					<div style={{ flex: 4 }}>
 						<TextField
 							multiline
-							style={{flex: "1", backgroundColor: inputColor}}
+							style={{ flex: "1", backgroundColor: theme.palette.inputColor }}
 							InputProps={{
-								style:{
+								style: {
 									color: "white",
 								},
 							}}
@@ -303,17 +305,17 @@ const Contact = (props) => {
 							id="filled-multiline-static"
 							margin="normal"
 							variant="outlined"
-      	 					onChange={e => setMessage(e.target.value)}
+							onChange={e => setMessage(e.target.value)}
 						/>
 					</div>
 					<Button
 						disabled={email.length <= 0 || message.length <= 0}
-						style={{width: "100%", height: "60px", marginTop: "10px"}}
+						style={{ width: "100%", height: "60px", marginTop: "10px" }}
 						variant="contained"
 						color="primary"
 						onClick={submitContact}
 					>
-					Submit	
+						Submit
 					</Button>
 					<h3>{formMessage}</h3>
 				</Paper>
@@ -321,7 +323,7 @@ const Contact = (props) => {
 		</div>
 
 
-	const loadedCheck = isLoaded ? 
+	const loadedCheck = isLoaded ?
 		<div>
 			<BrowserView>
 				<div style={bodyDivStyle}>{landingpageDataBrowser}</div>
@@ -334,7 +336,7 @@ const Contact = (props) => {
 		<div>
 		</div>
 
-	return(
+	return (
 		<div>
 			{loadedCheck}
 		</div>
