+++ frontend/src/views/LoginPage.jsx
@@ -1,13 +1,15 @@
 /* eslint-disable react/no-multi-comp */
-import React, {useState} from 'react';
+import React, { useState } from 'react';
 import { makeStyles } from '@material-ui/styles';
 
 import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
 import Paper from '@material-ui/core/Paper';
+import { useTheme } from '@material-ui/core/styles';
+
 
 const hrefStyle = {
-	color: "white", 
+	color: "white",
 	textDecoration: "none"
 }
 
@@ -17,16 +19,6 @@ const bodyDivStyle = {
 	width: "500px",
 }
 
-const surfaceColor = "#27292D"
-const inputColor = "#383B40"
-
-const boxStyle = {
-	paddingLeft: "30px",
-	paddingRight: "30px",
-	paddingBottom: "30px",
-	paddingTop: "30px",
-	backgroundColor: surfaceColor,
-}
 
 const useStyles = makeStyles({
 	notchedOutline: {
@@ -35,12 +27,14 @@ const useStyles = makeStyles({
 });
 
 const LoginDialog = props => {
-  const { globalUrl, isLoaded, isLoggedIn, setIsLoggedIn, setCookie, register } = props;
+	const theme = useTheme();
+
+	const { globalUrl, isLoaded, isLoggedIn, setIsLoggedIn, setCookie, register } = props;
 	const [username, setUsername] = useState("");
 	const [password, setPassword] = useState("");
 	const [firstRequest, setFirstRequest] = useState(true);
 
-  	// Used to swap from login to register. True = login, false = register
+	// Used to swap from login to register. True = login, false = register
 
 	const classes = useStyles();
 	// Error messages etc
@@ -51,32 +45,32 @@ const LoginDialog = props => {
 	}
 
 	if (isLoggedIn === true) {
-		window.location.pathname = "/workflows"		
+		window.location.pathname = "/workflows"
 	}
 
 
 	const checkAdmin = () => {
-		const url = globalUrl+'/api/v1/checkusers';
+		const url = globalUrl + '/api/v1/checkusers';
 		fetch(url, {
-				method: 'GET',
-				headers: {
-					'Content-Type': 'application/json',
-				},
+			method: 'GET',
+			headers: {
+				'Content-Type': 'application/json',
+			},
 		})
-		.then(response =>
-			response.json().then(responseJson => {
-				if (responseJson["success"] === false) {
-					setLoginInfo(responseJson["reason"])
-				} else {
-					if (responseJson.reason === "stay") {
-						window.location.pathname = "/adminsetup"		
+			.then(response =>
+				response.json().then(responseJson => {
+					if (responseJson["success"] === false) {
+						setLoginInfo(responseJson["reason"])
+					} else {
+						if (responseJson.reason === "stay") {
+							window.location.pathname = "/adminsetup"
+						}
 					}
-				}
-			}),
-		)
-		.catch(error => {
-			setLoginInfo("Error in userdata: ", error)
-		})
+				}),
+			)
+			.catch(error => {
+				setLoginInfo("Error in userdata: ", error)
+			})
 	}
 
 	if (firstRequest) {
@@ -89,41 +83,41 @@ const LoginDialog = props => {
 		// FIXME - add some check here ROFL
 
 		// Just use this one?
-		var data = {"username": username, "password": password}
+		var data = { "username": username, "password": password }
 		var baseurl = globalUrl
 		if (register) {
-			var url = baseurl+'/api/v1/users/login';
+			var url = baseurl + '/api/v1/users/login';
 			fetch(url, {
 				mode: 'cors',
-					method: 'POST',
-					body: JSON.stringify(data),
-					credentials: 'include',
-					crossDomain: true,
-					withCredentials: true,
-					headers: {
-        		'Content-Type': 'application/json; charset=utf-8',
-					},
+				method: 'POST',
+				body: JSON.stringify(data),
+				credentials: 'include',
+				crossDomain: true,
+				withCredentials: true,
+				headers: {
+					'Content-Type': 'application/json; charset=utf-8',
+				},
 			})
-			.then(response =>
-				response.json().then(responseJson => {
-					if (responseJson["success"] === false) {
-						setLoginInfo(responseJson["reason"])
-					} else {
-						setLoginInfo("Successful login, rerouting")
-						for (var key in responseJson["cookies"]) {
-							setCookie(responseJson["cookies"][key].key, responseJson["cookies"][key].value, {path: "/"})
-						}
+				.then(response =>
+					response.json().then(responseJson => {
+						if (responseJson["success"] === false) {
+							setLoginInfo(responseJson["reason"])
+						} else {
+							setLoginInfo("Successful login, rerouting")
+							for (var key in responseJson["cookies"]) {
+								setCookie(responseJson["cookies"][key].key, responseJson["cookies"][key].value, { path: "/" })
+							}
 
-						setIsLoggedIn(true)
-						window.location.pathname = "/workflows"		
-					}
-				}),
-			)
-			.catch(error => {
-				setLoginInfo("Error in userdata: " + error)
-			});
+							setIsLoggedIn(true)
+							window.location.pathname = "/workflows"
+						}
+					}),
+				)
+				.catch(error => {
+					setLoginInfo("Error in userdata: " + error)
+				});
 		} else {
-			url = baseurl+'/api/v1/users/register';
+			url = baseurl + '/api/v1/users/register';
 			fetch(url, {
 				method: 'POST',
 				body: JSON.stringify(data),
@@ -131,27 +125,27 @@ const LoginDialog = props => {
 					'Content-Type': 'application/json',
 				},
 			})
-			.then(response =>
-				response.json().then(responseJson => {
-					if (responseJson["success"] === false) {
-						setLoginInfo(responseJson["reason"])
-					} else {
-						setLoginInfo("Successful register :)")
-					}
-				}),
-			)
-			.catch(error => {
-				setLoginInfo("Error in userdata: ", error)
-			});
+				.then(response =>
+					response.json().then(responseJson => {
+						if (responseJson["success"] === false) {
+							setLoginInfo(responseJson["reason"])
+						} else {
+							setLoginInfo("Successful register :)")
+						}
+					}),
+				)
+				.catch(error => {
+					setLoginInfo("Error in userdata: ", error)
+				});
 		}
 	}
 
 	const onChangeUser = (e) => {
-		setUsername(e.target.value)	
+		setUsername(e.target.value)
 	}
 
 	const onChangePass = (e) => {
-		setPassword(e.target.value)	
+		setPassword(e.target.value)
 	}
 
 	//const onClickRegister = () => {
@@ -167,25 +161,31 @@ const LoginDialog = props => {
 	//var loginChange = register ? (<div><p onClick={setLoginCheck(false)}>Want to register? Click here.</p></div>) : (<div><p onClick={setLoginCheck(true)}>Go back to login? Click here.</p></div>);
 	var formtitle = register ? <div>Login</div> : <div>Register</div>
 
-		//	<DialogTitle>{formtitle}</DialogTitle>
-		
-	const basedata = 
+	//	<DialogTitle>{formtitle}</DialogTitle>
+
+	const basedata =
 		<div style={bodyDivStyle}>
-			<Paper style={boxStyle}>
-      	<form onSubmit={onSubmit} style={{margin: "15px 15px 15px 15px", color: "white",}}>
+			<Paper style={{
+				paddingLeft: "30px",
+				paddingRight: "30px",
+				paddingBottom: "30px",
+				paddingTop: "30px",
+				backgroundColor: theme.palette.surfaceColor,
+			}}>
+				<form onSubmit={onSubmit} style={{ margin: "15px 15px 15px 15px", color: "white", }}>
 					<h2>{formtitle}</h2>
 					Username
 					<div>
 						<TextField
 							color="primary"
-							style={{backgroundColor: inputColor}}
+							style={{ backgroundColor: theme.palette.inputColor }}
 							autoFocus
 							InputProps={{
 								classes: {
 									notchedOutline: classes.notchedOutline,
 								},
-								style:{
-									height: "50px", 
+								style: {
+									height: "50px",
 									color: "white",
 									fontSize: "1em",
 								},
@@ -197,20 +197,20 @@ const LoginDialog = props => {
 							id="emailfield"
 							margin="normal"
 							variant="outlined"
-      	 					onChange={onChangeUser}
+							onChange={onChangeUser}
 						/>
 					</div>
-					Password	
+					Password
 					<div>
 						<TextField
 							color="primary"
-							style={{backgroundColor: inputColor,}}
+							style={{ backgroundColor: theme.palette.inputColor }}
 							InputProps={{
 								classes: {
 									notchedOutline: classes.notchedOutline,
 								},
-								style:{
-									height: "50px", 
+								style: {
+									height: "50px",
 									color: "white",
 									fontSize: "1em",
 								},
@@ -223,23 +223,23 @@ const LoginDialog = props => {
 							placeholder="**********"
 							margin="normal"
 							variant="outlined"
-      	 					onChange={onChangePass}
+							onChange={onChangePass}
 						/>
 					</div>
-					<div style={{display: "flex", marginTop: "15px"}}>
-						<Button color="primary" variant="contained" type="submit" style={{flex: "1", marginRight: "5px"}} disabled={!handleValidateForm()}>SUBMIT</Button> 				
-						
+					<div style={{ display: "flex", marginTop: "15px" }}>
+						<Button color="primary" variant="contained" type="submit" style={{ flex: "1", marginRight: "5px" }} disabled={!handleValidateForm()}>SUBMIT</Button>
+
 					</div>
-					<div style={{marginTop: "10px"}}>
+					<div style={{ marginTop: "10px" }}>
 						{loginInfo}
 					</div>
 				</form>
 			</Paper>
 		</div>
 
-	const loadedCheck = isLoaded ? 
+	const loadedCheck = isLoaded ?
 		<div>
-      		{basedata}
+			{basedata}
 		</div>
 		:
 		<div>
