+++ frontend/src/views/PrivacyPolicy.jsx
