+++ frontend/src/views/Workflows.jsx
@@ -988,7 +988,7 @@ const Workflows = (props) => {
 					<div style={{flex: "1"}}>
 				    	<FormControlLabel
 							style={{color: "white", marginBottom: "0px", marginTop: "10px"}}
-							label=<div style={{color: "white"}}>Collapse results</div>
+							label={<div style={{color: "white"}}>Collapse results</div>}
 							control={<Switch checked={collapseJson} onChange={() => {setCollapseJson(!collapseJson)}} />}
 						/>
 					</div>
